Fix: added check of all methods in task19

+++ homework_3/task19.cs
@@ -27,7 +27,7 @@ namespace homework3 {
 
         }
 
-        public void isPolindromIntNumV1(int num) {
+        public void isPolindromIntNumV1(long num) {
             
             string number = Convert.ToString(num);
